Changed the type of the ::at method.
  Also added debug-assertions.

+++ history.h
@@ -44,7 +44,7 @@ class CHistory
   /**
    * Get the Nth piece of history.
    */
-  std::string at( int offset );
+  std::string at( size_t offset );
 
   /**
    * Add a new string to the history.
